Add L2C tracking

+++ src/sbp_utils.c
@@ -294,7 +294,7 @@ void unpack_ephemeris(const msg_ephemeris_t *msg, ephemeris_t *e)
   e->sid              = sid_from_sbp(msg->sid);
   e->kepler.iode      = msg->iode;
   e->kepler.iodc      = msg->iodc;
-  e->fit_interval     = 4 * 60 * 60; /* TODO: this is a work around until SBP updated */
+  e->fit_interval     = (u8)(4 * 60 * 60); /* TODO: this is a work around until SBP updated */
   e->ura              = 2.0f; /* TODO: this is a work around until SBP updated*/
 }
